+++ dendritex/_base.py
@@ -25,7 +25,7 @@ import numpy as np
 from brainstate.mixin import _JointGenericAlias
 
 from ._integrators import DiffEqModule
-from ._misc import set_module_as
+from ._misc import set_module_as, Container, TreeNode
 
 __all__ = [
     'HHTypedNeuron',
@@ -47,95 +47,6 @@ __all__ = [
   - MixIons
   - Channel
 '''
-
-
-class Container(bst.mixin.Mixin):
-    __module__ = 'dendritex'
-
-    _container_name: str
-
-    @staticmethod
-    def _format_elements(child_type: type, **children_as_dict):
-        res = dict()
-
-        # add dict-typed components
-        for k, v in children_as_dict.items():
-            if not isinstance(v, child_type):
-                raise TypeError(f'Should be instance of {child_type.__name__}. '
-                                f'But we got {type(v)}')
-            res[k] = v
-        return res
-
-    def __getitem__(self, item):
-        """
-        Overwrite the slice access (`self['']`).
-        """
-        children = self.__getattr__(self._container_name)
-        if item in children:
-            return children[item]
-        else:
-            raise ValueError(f'Unknown item {item}, we only found {list(children.keys())}')
-
-    def __getattr__(self, item):
-        """
-        Overwrite the dot access (`self.`).
-        """
-        name = super().__getattribute__('_container_name')
-        if item == '_container_name':
-            return name
-        children = super().__getattribute__(name)
-        if item == name:
-            return children
-        if item in children:
-            return children[item]
-        else:
-            return super().__getattribute__(item)
-
-    def add_elem(self, *elems, **elements):
-        """
-        Add new elements.
-
-        Args:
-          elements: children objects.
-        """
-        raise NotImplementedError('Must be implemented by the subclass.')
-
-
-class TreeNode(bst.mixin.Mixin):
-    __module__ = 'dendritex'
-
-    root_type: type
-
-    @staticmethod
-    def _root_leaf_pair_check(root: type, leaf: 'TreeNode'):
-        if hasattr(leaf, 'root_type'):
-            root_type = leaf.root_type
-        else:
-            raise ValueError('Child class should define "root_type" to '
-                             'specify the type of the root node. '
-                             f'But we did not found it in {leaf}')
-        if not issubclass(root, root_type):
-            raise TypeError(f'Type does not match. {leaf} requires a root with type '
-                            f'of {leaf.root_type}, but the root now is {root}.')
-
-    @staticmethod
-    def check_hierarchies(root: type, *leaves, check_fun: Callable = None, **named_leaves):
-        if check_fun is None:
-            check_fun = TreeNode._root_leaf_pair_check
-
-        for leaf in leaves:
-            if isinstance(leaf, bst.graph.Node):
-                check_fun(root, leaf)
-            elif isinstance(leaf, (list, tuple)):
-                TreeNode.check_hierarchies(root, *leaf, check_fun=check_fun)
-            elif isinstance(leaf, dict):
-                TreeNode.check_hierarchies(root, **leaf, check_fun=check_fun)
-            else:
-                raise ValueError(f'Do not support {type(leaf)}.')
-        for leaf in named_leaves.values():
-            if not isinstance(leaf, bst.graph.Node):
-                raise ValueError(f'Do not support {type(leaf)}. Must be instance of {bst.graph.Node}')
-            check_fun(root, leaf)
 
 
 class HHTypedNeuron(bst.nn.Dynamics, Container, DiffEqModule):
@@ -193,13 +104,13 @@ class HHTypedNeuron(bst.nn.Dynamics, Container, DiffEqModule):
         return self.update(*args, **kwargs)
 
     def init_state(self, batch_size=None):
-        nodes = self.nodes(allowed_hierarchy=(1, 1)).filter(IonChannel).values()
+        nodes = self.nodes(IonChannel, allowed_hierarchy=(1, 1)).values()
         TreeNode.check_hierarchies(self.__class__, *nodes)
         for channel in nodes:
             channel.init_state(self.V.value, batch_size=batch_size)
 
     def reset_state(self, batch_size=None):
-        nodes = self.nodes(allowed_hierarchy=(1, 1)).filter(IonChannel).values()
+        nodes = self.nodes(IonChannel, allowed_hierarchy=(1, 1)).values()
         for channel in nodes:
             channel.reset_state(self.V.value, batch_size=batch_size)
 
@@ -312,7 +223,7 @@ class Ion(IonChannel, Container):
         name: Optional[str] = None,
         **channels
     ):
-        super().__init__(size, name=name, **channels)
+        super().__init__(size, name=name)
         self.channels: Dict[str, Channel] = dict()
         self.channels.update(self._format_elements(Channel, **channels))
